pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imagen-apex"
version = "1.0.0"
description = "Text-to-3D generation pipeline using Nano Banana Pro and SAM 3D"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Luke Hamond"}
]
requires-python = ">=3.11"
keywords = ["3d", "ai", "image-generation", "text-to-3d", "vertex-ai", "sam3d"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
]
dependencies = [
    "google-cloud-aiplatform>=1.38.0",
    "vertexai>=1.0.0",
    "google-cloud-storage>=2.10.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

[project.scripts]
imagen-apex = "src.pipeline:main"
imagen-generate = "src.image_generator:main"

[project.urls]
Homepage = "https://github.com/luke-hamond/imagen-apex"
Repository = "https://github.com/luke-hamond/imagen-apex.git"
Issues = "https://github.com/luke-hamond/imagen-apex/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
